requirements.txt
Fabric==1.5.3
Jinja2==2.6
Markdown==2.2.1
PyYAML==3.10
Pygments==1.6
argparse==1.2.1
commando==0.3a
hyde==0.8.4
hydewiki==0.1
mediawiki==0.0.1
mobileplugin==0.2
paramiko==1.9.0
pycrypto==2.6
rmll-hyde-extensions==0.1
smartypants==1.6.0.3
typogrify-hyde==1.0.1a
wsgiref==0.1.2
